Tests/LibWeb/Text/expected/geolocation-clear-watch-emulated.txt

update 1: 37.7647658, -122.4345892
updates after clearWatch: 1
